# sentinelapi/vulnerable-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SentinelAPI Vulnerable Test Target",
    "description": "Intentionally vulnerable sample API designed for security assessment testing. Demonstrates BOLA/IDOR, excessive data exposure, and missing rate limiting.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local test sandbox"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer authentication token"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "alice"
          },
          "password": {
            "type": "string",
            "example": "alice123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "username": {
                "type": "string",
                "example": "alice"
              }
            }
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Mechanical Keyboard",
              "USB-C Hub"
            ]
          },
          "amount": {
            "type": "number",
            "example": 189.99
          },
          "status": {
            "type": "string",
            "example": "completed"
          },
          "card_number": {
            "type": "string",
            "example": "4532015698741235"
          },
          "created_at": {
            "type": "string",
            "example": "2026-09-20T10:15:00Z"
          },
          "payment_id": {
            "type": "integer",
            "example": 501
          },
          "invoice_id": {
            "type": "integer",
            "example": 901
          },
          "shipment_id": {
            "type": "integer",
            "example": 701
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "alice"
          },
          "email": {
            "type": "string",
            "example": "alice@example.com"
          },
          "password": {
            "type": "string",
            "example": "alice123"
          },
          "role": {
            "type": "string",
            "example": "customer"
          },
          "bio": {
            "type": "string",
            "example": "Senior developer and coffee enthusiast"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "internalNotes": {
            "type": "string"
          },
          "is_admin": {
            "type": "boolean"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Resource not found"
          },
          "status": {
            "type": "integer",
            "example": 404
          }
        }
      },
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "order_id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "amount": {
            "type": "number"
          },
          "card_number": {
            "type": "string"
          },
          "payment_token": {
            "type": "string"
          }
        }
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "order_id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "billing_address": {
            "type": "string"
          },
          "internal_metadata": {
            "type": "string"
          }
        }
      },
      "Shipment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "order_id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "tracking_number": {
            "type": "string"
          },
          "destination_address": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Authenticate user",
        "description": "Authenticates credentials and returns a Bearer access token. Intentionally lacks rate limiting.",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get order details by ID",
        "description": "VULNERABLE (BOLA & Excessive Data Exposure): Lacks ownership authorization checks. Any authenticated user can access any other user's order and views unmasked payment card details.",
        "operationId": "getOrderById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric Order ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get user profile by ID",
        "description": "VULNERABLE (BOLA & Credential Exposure): Any authenticated user can retrieve profiles of other users, which includes their plaintext password in the response body.",
        "operationId": "getUserById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "User ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update user profile",
        "operationId": "updateUser",
        "description": "Intentionally vulnerable mass-assignment endpoint; protected fields are writable.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string"
                  },
                  "is_admin": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/me/orders": {
      "get": {
        "summary": "Get current authenticated user's orders",
        "description": "SECURE PATTERN: Enforces ownership by deriving user identity strictly from the authenticated token context. Only returns orders belonging to the requester.",
        "operationId": "getMyOrders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of current user's orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "Get OpenAPI 3.0 specification",
        "description": "Serves this OpenAPI JSON document.",
        "operationId": "getOpenApiSpec",
        "responses": {
          "200": {
            "description": "OpenAPI specification document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/payment": {
      "get": {
        "summary": "Get order payment",
        "operationId": "getOrderPayment",
        "description": "Intentionally vulnerable nested resource route; authorization is not checked.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Demo response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/orders/{id}/invoice": {
      "get": {
        "summary": "Get order invoice",
        "operationId": "getOrderInvoice",
        "description": "Intentionally vulnerable nested resource route; authorization is not checked.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Demo response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/orders/{id}/shipment": {
      "get": {
        "summary": "Get order shipment",
        "operationId": "getOrderShipment",
        "description": "Intentionally vulnerable nested resource route; authorization is not checked.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Demo response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shipment"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/admin/users/{id}/role": {
      "post": {
        "summary": "Change a user role",
        "operationId": "changeUserRole",
        "description": "Intentionally vulnerable BFLA endpoint; no admin role check is applied.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role changed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  }
}